Correct change in heading using right mouse button.
The desired behavior is to have the globe follow the mouse movement, e.g.,
when the mouse is above mid-screen, dragging the mouse to the right should
cause a clockwise rotation of the globe.  The change from AWT to GL screen
coordinates had broken this.  Thanks to Eugene Maksymenko (WorldWindEarth).